third_party/WebKit/LayoutTests/imported/wpt/service-workers/service-worker/controller-on-disconnect.https-expected.txt

This is a testharness.js-based test.
FAIL controller is cleared on disconnected window assert_equals: disconnected frame should not be controlled expected null but got object "[object ServiceWorker]"
Harness: the test ran to completion.
